Add better workflow for tracking
The tracking is now started, when the user enters the
TrackingViewController, he can then navigate around the App without the
Tracking beeing stopped. This is realised by a singleton Tracker object.
If the user stops the tracking in the TrackingView, he can start a new
track afterwards from the RootViewController.