Use a NSOperationQueue to queue KeilerHTTPOperations
The KeilerClient now holds a queue to which it will add the
KeilerHTTPOperations it was only starting so far. By configuring the
queue to only execute one operation at a time and the fact that an
operation is not finished until the HTTP request finishs only one
coordinate can be uploaded and only after the previous finished
uploading.