Add delete function to Tracks that belong to a user
This commit adds a DELETEOperation and uncovers a problem with
authentication when not doing POST Requests. Data can't be passed by
setting HTTPBody in the request header. Instead it must added as query
string. The authentication token will thus always be added to the query
string as this works for POST and the other types as well.