Add StalkerService to stalk the user
This commit does change a lot of things regarding the tracking and
stalking logic. The stalker is now only able to stalk the current
position because the MKPolyline is just to complicated to be easy
extended. So either a finished track (finishing call yet to implement)
is displayed in a whole or a not-yet-finished track will just display
the last known position.