Add last seen line to UsersViewController table
The last seen date is green, when the users is online, blue when he is
offline and red if he has a nil date. The "online" property is
determined by the server and send as json response for the users
collection.

Add animation to gravatar loading and don't reload gravatars when the
email has not changed.